requirements.txt
apprise
facebook-scraper
python-daemon
requests
